+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## 0.1.0
+
+### Minor Changes
+
+- - Refactored 'Optimize' Context Menu to 'Image Optimizer' with the following:
+    1. Compress
+    2. Convert
+  - Added the following Sub Menus to 'Convert' submenu
+    1.  Convert to JPG
+    2.  Convert to WebP
+    3.  Convert ro PNG
+
 All notable changes to the "image-optimizer" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.